requirements.txt
langchain>=0.1.0
langchain-community>=0.0.10
langchain-google-genai>=1.0.0
langchain-groq>=0.1.0
chromadb>=0.4.22
streamlit>=1.42.0
streamlit-sortables>=0.3.1
Authlib>=1.3.2
google-auth>=2.0.0
pypdf2>=3.0.1
python-dotenv>=1.0.0
torch==2.6.0
sentence-transformers>=3.0.0
pillow>=10.0.0
pytesseract>=0.3.10
pdf2image>=1.16.3
langchain-text-splitters>=0.0.1
requests>=2.31.0
httpx>=0.27.0
rank-bm25>=0.2.2
pandas>=2.0.0
bcrypt>=4.0.0
python-docx>=1.0.0
fastapi>=0.115.0
uvicorn>=0.30.0
slowapi>=0.1.9
pytest>=8.0.0
itsdangerous>=2.0.0